Make event_loop() process all messages in the buffer
Currently, event_loop() processes only one message per iteration, even
if there are more in the receive buffer. This means that messages can
build up in the receive buffer (I have observed this happening during
the Particle Cloud handshake), which might be problematic.

It's not clear if this this is intentional or an oversight. The
overridden version of event_loop() which waits for a specific message
does loop. I don't see any negative consequences of processing
everything in the buffer, so this commit modifies event_loop() to do so.
It yields between messages and I have also added a corresponding yield
to the specific message event_loop().